File handling moved into a separate module
