+++ promis/functions/potential.py
@@ -30,151 +30,12 @@ import re
 from io import StringIO
 import util.orbit
 import backend_api.models as model
+import util.files
 
 # TODO: move somewhre?
 import datetime, pytz
 def fromtimestamp(u):
     return datetime.datetime.fromtimestamp(u, tz=pytz.utc)
-
-# TODO: replace ValueErrors with meaningful exception classes when integrating
-# or make something like DataImportError(), whatever
-
-def file_catalog(fp):
-    # We have a bit of a decision here:
-    # 1. Read the file in one try reading each compontent into an in memory list
-    # 2. Read one point per time, but seek all over the file
-    # I'm going with route #2 for now, no particular reason, just think it will be more elegant in code
-
-    # Pointers to the current reading positions in the file per section
-    sections_index = { "RX": -1 , "RY": -1, "RZ": -1 }
-
-    # Regular expression to match the sections above
-    sections_rx = "^@(%s)" % "|".join(sections_index)
-
-    # Populating the index
-    while True: # Unable to use tell() with for interatin
-        ln = fp.readline()
-        if ln == "":
-            break
-        m = re.search(sections_rx, ln)
-        if m:
-            sect = m.group(1)
-            if sections_index[sect] > 0:
-                raise ValueError("Duplicate section detected")
-            sections_index[sect] = fp.tell()
-
-    # Checking if we found all the sections
-    if any((pos < 0 for _,pos in sections_index.items())):
-        raise ValueError("Some sections missing from input")
-
-    def scan_sect(sect):
-        while True:
-            # Come back to the saved position
-            fp.seek(sections_index[sect])
-
-            # Read a line, exit the iteration if it's the end of input
-            ln = fp.readline()
-            if ln.rstrip() == "":
-                break
-
-            # Record the new position
-            sections_index[sect] = fp.tell()
-
-            # Try to parse the data
-            # NOTE: expected format: <timestamp> <float value(.)> <human readable date>, ignoring the last one
-            m = re.search("^([0-9]*) ([0-9.-]*)", ln)
-            if m:
-                # Yielding a nested tuple e.g. ( "RX", (1, 432.0) ), will be converted to dict
-                yield ( sect, (int(m.group(1)), float(m.group(2))) )
-            else:
-                raise ValueError("Input inconsistency detected")
-
-    def scan_point():
-        # Generator objects for all the sections
-        sect_gens = [scan_sect(sect) for sect in sections_index]
-        try:
-            while True:
-                # Try reading the next point into a list, check if times are consistent
-                nextpoint = [next(g) for g in sect_gens]
-                timemark = nextpoint[0][1][0] # may throw an exception if len(nextpoint) == 0 for some reason
-
-                if any(( pt[1][0] != timemark for pt in nextpoint )):
-                    raise ValueError("Timemarks not consistent across sections evaluated")
-
-                # Convert to dictionary and remove redundant timemarks
-                nextpoint = dict(nextpoint)
-                for k,v in nextpoint.items():
-                    nextpoint[k] = v[1]
-
-                yield timemark, util.orbit.cord_conv(**nextpoint)
-        except StopIteration:
-            pass
-
-        # TODO: for super consistency, try advacing the rest of generators and
-        # make sure all of them raise StopIteration or something?
-
-    # Call the machinery above
-    for t, pt in scan_point():
-        yield (t, pt)
-
-def setfile_vars(fp, keys=None):
-    """
-    Look up key=value pairs in a file.
-
-    Value is assumed to be a number. If the key appears multiple times, only the first occurrence is returned.
-    If keys is set to a set of strings, only yield keys from it.
-    """
-    keys_found = set()
-    keys_left = -1 if not keys else len(keys)
-    # TODO: currently only integers, see regex below and yield statement
-
-    for m in re.finditer("([a-zA-Z_]+)=([0-9]+)", fp.getvalue()): # TODO: matches " = ", shouldn't ideally
-        key = m.group(1)
-        value = m.group(2)
-
-        # Skip entries we've seen before
-        # Skip entries we do not want
-        # Temporary measure: skip empty entries
-        if key in keys_found or (keys and key not in keys) or len(key) == 0 or len(value) == 0:
-            continue
-
-        # Register the key as found
-        keys_found.add(key)
-
-        # Yield the data
-        yield key, int(value)
-
-        # Reduce the counter of keys to look for and break if necessary
-        if keys_left > 0:
-            keys_left -= 1
-            if keys_left == 0:
-                break
-
-# TODO: generalise!
-def csvfile_1st_column(fp, as_type=int):
-    """
-    Skips comments in a .csv file and yields values of the first column.
-
-    fp      -- file descriptor.
-    as_type -- type to convert to.
-    """
-    rexp_comment = r"^#"
-    rexp_values  = r"^([0-9-.])*," # NOTE: numbers only
-
-    while True:
-        ln = fp.readline()
-        if ln.rstrip() == "":
-            break
-
-        # Skipping the comments
-        # TODO: anything of value here?
-        if re.search(rexp_comment, ln):
-            continue
-
-        # Trying to extract the value
-        m = re.search(rexp_values, ln)
-        if m:
-            yield as_type(m.group(1))
 
 def guess_duration(n, freq):
     """
@@ -244,7 +105,7 @@ def data_func(satellite_object):
                     # Retrieving and processing the raw file
                     ftp.retrlines("RETR " + fname, lambda x: fp.write(x + "\n"))
                     fp.seek(0)
-                    rawdata = dict(pt for pt in file_catalog(fp))
+                    rawdata = { t:pt for t, pt in util.files.telemetry(fp) }
 
                     # Append the data, assuming no repetitions can happen
                     orbit.update(rawdata)
@@ -298,7 +159,7 @@ def data_func(satellite_object):
                         with StringIO() as fp:
                             ftp.retrlines("RETR " + mvfile[0], lambda x: fp.write(x + "\n"))
                             fp.seek(0)
-                            data = { k:v for k,v in setfile_vars(fp, {"t", "samp"}) }
+                            data = { k:v for k,v in util.files.sets(fp, {"t", "samp"}) }
                             time_start = data["t"]
                             time_end = data["t"] + guess_duration(data["samp"], freqs[freq])
 
@@ -334,7 +195,7 @@ def data_func(satellite_object):
                             fp.seek(0)
                             
                             # Creating the JSON document
-                            mv = [ i for i in csvfile_1st_column(fp) ]
+                            mv = [ i for i in util.files.csv(fp) ]
                             # TODO: discuss the meaning of last_mod in details
                             doc_obj = model.Document.objects.create(json_data = { "mv": mv } )
                             
@@ -349,6 +210,5 @@ def data_func(satellite_object):
                     ftp.cwd("..")
 
                 ftp.cwd("..")
-
 
     return check, fetch
